Extract MOB_INDEX_DATA
