+++ src/db2.cpp
@@ -10,6 +10,7 @@
 #include "AFFECT_DATA.hpp"
 #include "AREA_DATA.hpp"
 #include "Logging.hpp"
+#include "MOB_INDEX_DATA.hpp"
 #include "db.h"
 #include "lookup.h"
 #include "merc.h"
