+++ src/merc.h
@@ -61,10 +61,10 @@ struct OBJ_INDEX_DATA;
 typedef struct program PROGRAM;
 typedef struct reset_data RESET_DATA;
 struct ROOM_INDEX_DATA;
-typedef struct shop_data SHOP_DATA;
+struct SHOP_DATA;
 typedef struct known_players KNOWN_PLAYERS; // TODO(#108) remove if unused.
 /* Merc22 MOBProgs */
-typedef struct mob_prog_data MPROG_DATA; /* MOBprogram */
+struct MPROG_DATA; /* MOBprogram */
 
 /*
  * Site ban structure.
@@ -135,7 +135,7 @@ struct con_app_type {
  */
 #define MAX_TRADE 5
 
-struct shop_data {
+struct SHOP_DATA {
     SHOP_DATA *next; /* Next shop in list            */
     sh_int keeper; /* Vnum of shop keeper mob      */
     sh_int buy_type[MAX_TRADE]; /* Item types shop will buy     */
@@ -927,51 +927,6 @@ static inline constexpr auto ff = BIT(31);
 #define COMM_NOALLEGE (X)
 
 /*
- * Prototype for a mob.
- * This is the in-memory version of #MOBILES.
- */
-struct MOB_INDEX_DATA {
-    MOB_INDEX_DATA *next;
-    SpecialFunc spec_fun;
-    SHOP_DATA *pShop;
-    sh_int vnum;
-    sh_int count;
-    sh_int killed;
-    std::string player_name;
-    std::string short_descr;
-    std::string long_descr;
-    std::string description;
-    unsigned long act;
-    unsigned long affected_by;
-    sh_int alignment;
-    sh_int group; /* rom-2.4 style mob groupings */
-    sh_int level;
-    sh_int hitroll;
-    sh_int hit[3];
-    sh_int mana[3];
-    sh_int damage[3];
-    sh_int ac[4];
-    sh_int dam_type;
-    long off_flags;
-    long imm_flags;
-    long res_flags;
-    long vuln_flags;
-    sh_int start_pos;
-    sh_int default_pos;
-    sh_int sex;
-    sh_int race;
-    long gold;
-    long form;
-    long parts;
-    sh_int size;
-    sh_int material;
-    MPROG_DATA *mobprogs; /* Used by MOBprogram */
-    int progtypes; /* Used by MOBprogram */
-
-    AREA_DATA *area;
-};
-
-/*
  * MOBprogram block
  */
 
@@ -983,7 +938,7 @@ struct MPROG_ACT_LIST {
     const void *vo;
 };
 
-struct mob_prog_data {
+struct MPROG_DATA {
     MPROG_DATA *next;
     AREA_DATA *area;
     int type;
